# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/churches:
    get:
      operationId: getAdminChurches
      tags: [app-admin]
      summary: List churches for the platform owner
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EarlyAccessStatus"
        - name: billingPlan
          in: query
          required: false
          schema:
            type: string
            enum: [starter, growing, complete, network, unlimited]
      responses:
        "200":
          description: Church management list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminChurchList"
        "401":
          description: Unauthorized
        "403":
          description: App administrator access required
  /admin/churches/{id}:
    get:
      operationId: getAdminChurch
      tags: [app-admin]
      summary: Get one church for the platform owner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Church management detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminChurchDetail"
        "401":
          description: Unauthorized
        "403":
          description: App administrator access required
        "404":
          description: Church not found
    patch:
      operationId: updateAdminChurch
      tags: [app-admin]
      summary: Update safe church management fields for the platform owner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminChurchUpdateInput"
      responses:
        "200":
          description: Updated church management detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminChurchDetail"
        "400":
          description: Invalid church update
        "401":
          description: Unauthorized
        "403":
          description: App administrator access required
        "404":
          description: Church not found
...
  /church:
    get:
      operationId: getMyChurch
      tags: [churches]
      summary: Get the signed-in administrator's church
      responses:
        "200":
          description: Church
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Church"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMyChurch
      tags: [churches]
      summary: Update church setup information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChurchUpdate"
      responses:
        "200":
          description: Updated church
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Church"
        "401":
          description: Unauthorized
  /church/admins:
    get:
      operationId: listChurchAdmins
      tags: [churches]
      summary: List pastors with access to the signed-in church
      responses:
        "200":
          description: Church administrators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChurchAdmin"
        "401":
          description: Unauthorized
    post:
      operationId: addChurchAdmin
      tags: [churches]
      summary: Add an existing Clerk user as a church administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChurchAdminAddInput"
      responses:
        "201":
          description: Added administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChurchAdmin"
        "400":
          description: Invalid administrator
        "401":
          description: Unauthorized
        "404":
          description: No Clerk account was found for the email address
        "409":
          description: User is already an administrator
  /church/admins/{id}:
    delete:
      operationId: removeChurchAdmin
      tags: [churches]
      summary: Remove a pastor's access to the signed-in church
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Administrator removed
        "400":
          description: The church owner cannot be removed
        "401":
          description: Unauthorized
        "404":
          description: Administrator not found
  /churches/{slug}:
    get:
      operationId: getPublicChurch
      tags: [churches]
      summary: Get the public church profile destination
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public church information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicChurch"
        "404":
          description: Church not found
        "422":
          description: Church assessment configuration is invalid
  /churches/{slug}/admin-access:
    get:
      operationId: getChurchAdminAccess
      tags: [churches]
      summary: Check whether the signed-in user may override a youth pathway
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Church-specific override access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChurchAdminAccess"
        "401":
          description: Unauthorized
        "404":
          description: Church not found
  /churches/{slug}/discover-access:
    post:
      operationId: verifyDiscoverAccess
      tags: [churches]
      summary: Verify a church's current Discover hallway code
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 120
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscoverAccessInput"
      responses:
        "200":
          description: Discover access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverAccessResponse"
        "401":
          description: Missing or incorrect hallway code
        "404":
          description: Church not found
  /churches/{slug}/logo:
    get:
      operationId: getPublicChurchLogo
      tags: [churches]
      summary: Get the approved public logo for a church
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Church logo image
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: Logo not found
...
components:
  schemas:
...
    EarlyAccessStatus:
      type: string
      enum: [early_access, standard]
...
    AdminChurchSummary:
      type: object
      required: [id, name, slug, address, website, adminName, adminEmail, earlyAccessStatus, earlyAccessStartDate, foundingChurch, billingPlan, billingStatus, billingCurrentPeriodEnd, completedProfileCount, activePeopleCount, adminCount, lastActivityAt, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        address: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        adminName: { type: string }
        adminEmail: { type: string, format: email }
        earlyAccessStatus:
          $ref: "#/components/schemas/EarlyAccessStatus"
        earlyAccessStartDate: { type: string, format: date-time }
        foundingChurch: { type: boolean }
        billingPlan:
          type: string
          enum: [starter, growing, complete, network, unlimited]
        billingStatus: { type: string }
        billingCurrentPeriodEnd:
          type: ["string", "null"]
          format: date-time
        completedProfileCount: { type: integer }
        activePeopleCount: { type: integer }
        adminCount: { type: integer }
        lastActivityAt:
          type: ["string", "null"]
          format: date-time
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AdminChurchList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminChurchSummary"
    AdminChurchAdmin:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        role:
          type: string
          enum: [owner, admin]
        createdAt: { type: string, format: date-time }
    AdminChurchDetail:
      type: object
      required: [church, admins]
      properties:
        church:
          $ref: "#/components/schemas/AdminChurchSummary"
        admins:
          type: array
          items:
            $ref: "#/components/schemas/AdminChurchAdmin"
    AdminChurchUpdateInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        address: { type: ["string", "null"], maxLength: 500 }
        website: { type: ["string", "null"], maxLength: 500 }
        adminName: { type: string, minLength: 1, maxLength: 255 }
        adminEmail: { type: string, format: email }
        earlyAccessStatus:
          $ref: "#/components/schemas/EarlyAccessStatus"
        foundingChurch: { type: boolean }
...
    Church:
      type: object
      required: [id, name, slug, profileUrl, adminName, adminEmail, primaryColor, accentColor, assessmentConfiguration, ministryCustomization, onboardingCompletedAt, earlyAccessStatus, earlyAccessStartDate, foundingChurch, billingPlan, billingStatus, billingCurrentPeriodEnd, discoverHallwayCode]
      properties:
        integratedAssessmentPilotEnabled: { type: boolean, default: false }
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        logoUrl: { type: ["string", "null"] }
        primaryColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        accentColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        address: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        discoverHallwayCode: { type: ["string", "null"], pattern: "^[A-HJ-NP-Z2-9]{6}$" }
        adminName: { type: string }
        adminEmail: { type: string, format: email }
        profileUrl: { type: string }
        completedProfileCount: { type: integer }
        enabledSpiritualGifts:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/SpiritualGiftName"
          uniqueItems: true
        assessmentConfiguration:
          $ref: "#/components/schemas/AssessmentConfiguration"
        ministryCustomization:
          $ref: "#/components/schemas/MinistryCustomization"
        onboardingCompletedAt:
          type: ["string", "null"]
          format: date-time
        earlyAccessStatus:
          $ref: "#/components/schemas/EarlyAccessStatus"
        earlyAccessStartDate:
          type: string
          format: date-time
        foundingChurch: { type: boolean }
        billingPlan:
          type: string
          enum: [starter, growing, complete, network, unlimited]
        billingStatus: { type: string }
        billingCurrentPeriodEnd:
          type: ["string", "null"]
          format: date-time
...
    ChurchUpdate:
      type: object
      properties:
        integratedAssessmentPilotEnabled: { type: boolean }
        name: { type: string, minLength: 1 }
        logoUrl:
          type: ["string", "null"]
          pattern: "^/objects/logos/[0-9]+/[0-9a-fA-F-]+$"
        primaryColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        accentColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        address: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        discoverHallwayCode: { type: string, minLength: 6, maxLength: 6, pattern: "^[A-HJ-NP-Z2-9]{6}$" }
        adminName: { type: string }
        adminEmail: { type: string, format: email }
        enabledSpiritualGifts:
          type: array
          minItems: 3
          uniqueItems: true
          items:
            $ref: "#/components/schemas/SpiritualGiftName"
        assessmentConfiguration:
          $ref: "#/components/schemas/AssessmentConfiguration"
        ministryCustomization:
          $ref: "#/components/schemas/MinistryCustomization"
        onboardingCompleted:
          type: boolean
    ChurchAdmin:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        role:
          type: string
          enum: [owner, admin]
        createdAt: { type: string, format: date-time }
    ChurchAdminAddInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ChurchAdminAccess:
      type: object
      additionalProperties: false
      required: [canOverrideYouthPathway]
      properties:
        canOverrideYouthPathway: { type: boolean }
    DiscoverAccessInput:
      type: object
      additionalProperties: false
      required: [hallwayCode]
      properties:
        hallwayCode:
          type: string
          minLength: 6
          maxLength: 6
          pattern: "^[A-HJ-NP-Z2-9]{6}$"
    DiscoverAccessResponse:
      type: object
      additionalProperties: false
      required: [authorized]
      properties:
        authorized: { type: boolean, enum: [true] }
    PublicChurch:
      type: object
      required: [name, slug, profileUrl, enabledSpiritualGifts, primaryColor, accentColor, assessmentConfiguration, ministryCustomization, discoverAccessAvailable]
      properties:
        id: { type: integer }
        integratedAssessmentPilotEnabled: { type: boolean, default: false }
        name: { type: string }
        slug: { type: string }
        logoUrl: { type: ["string", "null"] }
        primaryColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        accentColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        profileUrl: { type: string }
        discoverAccessAvailable:
          type: boolean
          description: Whether this church has enabled code-based access to Discover. The code itself is never public.
        enabledSpiritualGifts:
          type: array
          minItems: 3
          uniqueItems: true
          items:
            $ref: "#/components/schemas/SpiritualGiftName"
        assessmentConfiguration:
          $ref: "#/components/schemas/AssessmentConfiguration"
        ministryCustomization:
          $ref: "#/components/schemas/MinistryCustomization"
    MinistryCustomization:
      type: object
      additionalProperties: false
      required:
        - version
        - mode
        - tradition
        - customTradition
        - spiritualGiftsLabel
        - ministryInterestsLabel
      properties:
        version:
          type: integer
          enum: [1]
        mode:
          type: string
          enum: [standard, tradition, custom]
        tradition:
          type: string
          enum:
            - wesleyanHoliness
            - nazarene
            - methodist
            - baptist
            - pentecostalCharismatic
            - assembliesOfGod
            - presbyterianReformed
            - lutheran
            - anglicanEpiscopal
            - catholic
            - easternOrthodox
            - nonDenominational
            - independentEvangelical
            - other
            - preferNotToSpecify
        customTradition:
          type: ["string", "null"]
          maxLength: 120
        spiritualGiftsLabel:
          type: string
          minLength: 1
          maxLength: 80
        ministryInterestsLabel:
          type: string
          minLength: 1
          maxLength: 80
...
    SpiritualGiftName:
      type: string
      enum:
        - Administration
        - Apostleship
        - Discernment of Spirits
        - Evangelism
        - Exhortation / Encouragement
        - Faith
        - Giving
        - Healing
        - Helps / Service
        - Hospitality
        - Interpretation of Tongues
        - Knowledge
        - Leadership
        - Mercy
        - Miracles
        - Pastoring / Shepherding
        - Prophecy
        - Teaching
        - Tongues
        - Wisdom
        - Craftsmanship
        - Intercession
        - Missionary / Cross-Cultural Ministry
        - Music / Worship
        - Celibacy
        - Voluntary Poverty
...
    AssessmentConfiguration:
      type: object
      required: [sections, subsections, spiritualGiftQuestionCount, ministryQuestionCount, passions, ministryInterests]
      additionalProperties: false
      properties:
        sections:
          $ref: "#/components/schemas/AssessmentConfigurationSections"
        subsections:
          $ref: "#/components/schemas/AssessmentConfigurationSubsections"
        spiritualGiftQuestionCount:
          type: integer
          minimum: 1
          maximum: 4
          default: 3
          description: Number of reflection questions asked for each enabled spiritual gift in new adult assessments.
        ministryQuestionCount:
          type: integer
          minimum: 1
          maximum: 4
          default: 3
          description: Number of How You Minister reflection questions asked for each enabled ministry approach in new adult assessments.
        passions:
          type: array
          maxItems: 100
          items: { type: string, minLength: 1, maxLength: 80 }
        ministryInterests:
          type: array
          maxItems: 100
          items: { type: string, minLength: 1, maxLength: 80 }
        youthProfiles:
          $ref: "#/components/schemas/YouthProfilesConfiguration"
    YouthProfilesConfiguration:
      type: object
      additionalProperties: false
      required: [version, discover, explore, develop]
      properties:
        version: { type: integer, enum: [1] }
        discover: { $ref: "#/components/schemas/YouthProfileConfiguration" }
        explore: { $ref: "#/components/schemas/YouthProfileConfiguration" }
        develop: { $ref: "#/components/schemas/YouthProfileConfiguration" }
    YouthProfileConfiguration:
      type: object
      additionalProperties: false
      required: [profileTitle, profileDescription, sections, choiceLabels]
      properties:
        profileTitle: { type: string, minLength: 1, maxLength: 80 }
        profileDescription: { type: string, minLength: 1, maxLength: 300 }
        sections:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/YouthProfileSection"
        choiceLabels:
          type: object
          additionalProperties:
            type: string
            minLength: 1
            maxLength: 120
    YouthProfileSection:
      type: object
      additionalProperties: false
      required: [enabled, title, description]
      properties:
        enabled: { type: boolean }
        title: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, minLength: 1, maxLength: 300 }
    AssessmentConfigurationSections:
      type: object
      additionalProperties: false
      required: [aboutYou, apest, spiritualGifts, passionsInterests, naturalStrengths, personalityStrengths, spiritualHealth, connectionAvailability]
      properties:
        aboutYou: { type: boolean }
        apest: { type: boolean }
        spiritualGifts: { type: boolean }
        passionsInterests: { type: boolean }
        naturalStrengths: { type: boolean }
        personalityStrengths: { type: boolean }
        spiritualHealth: { type: boolean }
        connectionAvailability: { type: boolean }
...