# lib/api-zod/src/generated/types/churchUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentConfiguration } from './assessmentConfiguration';
import type { MinistryCustomization } from './ministryCustomization';
import type { SpiritualGiftName } from './spiritualGiftName';

export interface ChurchUpdate {
  integratedAssessmentPilotEnabled?: boolean;
  /** @minLength 1 */
  name?: string;
  /**
     * @nullable
     * @pattern ^/objects/logos/[0-9]+/[0-9a-fA-F-]+$
     */
  logoUrl?: string | null;
  /** @pattern ^#[0-9A-Fa-f]{6}$ */
  primaryColor?: string;
  /** @pattern ^#[0-9A-Fa-f]{6}$ */
  accentColor?: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  website?: string | null;
  /**
     * @minLength 6
     * @maxLength 6
     * @pattern ^[A-HJ-NP-Z2-9]{6}$
     */
  discoverHallwayCode?: string;
  adminName?: string;
  adminEmail?: string;
  /** @minItems 3 */
  enabledSpiritualGifts?: SpiritualGiftName[];
  assessmentConfiguration?: AssessmentConfiguration;
  ministryCustomization?: MinistryCustomization;
  onboardingCompleted?: boolean;
}